+++ src/Modifiers/PrependSegment.php
@@ -28,7 +28,7 @@ class PrependSegment extends AbstractPathModifier
     /**
      * New instance
      *
-     * @param Path|string $segment
+     * @param string $segment
      */
     public function __construct($segment)
     {
